requirements.txt
langchain==0.3.27
langchain-anthropic==0.3.19
langchain-openai==0.3.32
langchain-community==0.3.29
python-dotenv==1.1.1
pydantic==2.11.7
duckduckgo-search==8.1.1
wikipedia==1.4.0